requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask==2.2.5
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.7
requests==2.31.0
urllib3==2.0.4
Werkzeug==2.2.3
gunicorn==21.2.0
packaging==23.1
aniso8601==9.0.1
attrs==23.1.0
configparser==5.3.0
flask-restx==1.1.0
jsonschema==4.17.3
jsonschema-specifications==0.1.3
pytz==2023.3
referencing==0.8.9

# VERSION 1.2
cffi==1.15.1
cryptography==41.0.3
dotenv==0.0.5
Flask-Cors==4.0.0
funcs==0.8.0
named==1.3.0
pycparser==2.21
PyJWT==1.7.1
python-dotenv==1.0.0
rpds-py==0.10.0
solus==1.1.0
typing-aliases==1.1.3
typing_extensions==4.7.1
wraps==0.6.1
